Make spinlocks mandatory
With the change to make compare and swap mandatory, we can always
implement spinlocks, so make them mandatory and simplify the interface.
