+++ opal/include/opal/sys/arm64/atomic.h
@@ -257,6 +257,15 @@ static inline int64_t opal_atomic_swap_64(opal_atomic_int64_t *addr, int64_t new
 
 /**********************************************************************
  *
+ * Atomic spinlocks
+ *
+ *********************************************************************/
+
+#include "opal/sys/atomic_impl_spinlock.h"
+
+
+/**********************************************************************
+ *
  * Atomic math operations
  *
  *********************************************************************/
