+++ opal/include/opal/sys/atomic_impl.h
@@ -321,44 +321,3 @@ static inline intptr_t opal_atomic_sub_fetch_ptr(opal_atomic_intptr_t *addr, voi
     return 0;
 #    endif
 }
-
-
-/**********************************************************************
- *
- * Atomic spinlocks
- *
- *********************************************************************/
-#ifdef OPAL_NEED_INLINE_ATOMIC_SPINLOCKS
-
-/*
- * Lock initialization function. It set the lock to UNLOCKED.
- */
-static inline void opal_atomic_lock_init(opal_atomic_lock_t *lock, int32_t value)
-{
-    lock->u.lock = value;
-}
-
-static inline int opal_atomic_trylock(opal_atomic_lock_t *lock)
-{
-    int32_t unlocked = OPAL_ATOMIC_LOCK_UNLOCKED;
-    bool ret = opal_atomic_compare_exchange_strong_acq_32(&lock->u.lock, &unlocked,
-                                                          OPAL_ATOMIC_LOCK_LOCKED);
-    return (ret == false) ? 1 : 0;
-}
-
-static inline void opal_atomic_lock(opal_atomic_lock_t *lock)
-{
-    while (opal_atomic_trylock(lock)) {
-        while (lock->u.lock == OPAL_ATOMIC_LOCK_LOCKED) {
-            /* spin */;
-        }
-    }
-}
-
-static inline void opal_atomic_unlock(opal_atomic_lock_t *lock)
-{
-    opal_atomic_wmb();
-    lock->u.lock = OPAL_ATOMIC_LOCK_UNLOCKED;
-}
-
-#endif /* OPAL_HAVE_ATOMIC_SPINLOCKS */
