+++ opal/include/opal/sys/x86_64/atomic.h
@@ -152,6 +152,15 @@ static inline int64_t opal_atomic_swap_64(opal_atomic_int64_t *addr, int64_t new
 
 /**********************************************************************
  *
+ * Atomic spinlocks
+ *
+ *********************************************************************/
+
+#include "opal/sys/atomic_impl_spinlock.h"
+
+
+/**********************************************************************
+ *
  * Atomic math operations
  *
  *********************************************************************/
